New optional operator function, check(), it takes the same arguments as execute().  This runs after changing a property and allows correcting incompatible options.  Returning True will redraw the UI.
Currently this is used for setting the write extension when saving files, so changing the image format also corrects the extension.
The same is accessible from python where its used when saving SVG/EPS/PNG files.

This fixes: [#23828] obj export problems, [#23760] Exporting OBJ and filetype ending
also fixed document submission operator.


Now the filename in the file selector is the one used for writing this means we remove the "Save Over" popup which could be overlooked too easily.
Instead display the filename field with red tint, and a note in the tooltip.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@31978 954f8c5b-7b00-dc11-b283-0030488c597c