Fix [#23977] toggle back to object mode not working (outliner issue) Reported by Roland Kramer
There was already code to prevent visibility toggle through restrict column from working when in edit mode. Reshuffled
code somewhat so it works also for object operations in outliner. Also ensure operator poll for visibility and selectability toggle
checks object is not in edit mode. So this also works for selectability toggling, so no more toggling when in edit mode - it's confusing otherwise.

Added notifier and handling for it for renderability toggle in outliner. No edit mode restriction here.



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@32159 954f8c5b-7b00-dc11-b283-0030488c597c