fix for fix [#24336] evaluation time - curve has no effect committed r32598 to fix [#24309] Reloading file with incorrect path location.
Setting the curves ctime when updating the object data overrode the animated value.
For now just set the ctime on newly added curves with will work as the user expects in most cases.
This is weak design IMHO because the ctime value can be set to anything but is reset on changing frames even if its not keyed.

With curves created via python or linked in this can still result in a bad ctime value.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@32614 954f8c5b-7b00-dc11-b283-0030488c597c