Fix #26781: Assigning multiple Materials to Text seems not to work correctly
Problems was in 0-based numbering vs. 1-based. Fonts are getting converted to
displists when they're displaying and here base conversion happens, so no need
in conversion when assigning material to selected text.

This makes indices more inconsistant, but saves compatibility in both of
directions. Maybe in further releases we'll make indices consistent.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@36003 954f8c5b-7b00-dc11-b283-0030488c597c