Bugfix [#27194] Difficulty selecting bones with boneshapes enabled and some bones made un-selectable
Pose Bone selection (like object selection) used OpenGL to draw
wireframes with different colours into an offscreen buffer and then
from that identify which item was clicked on. The bug here was that
unselectable bones were getting drawn for this step too, so they were
getting caught instead of the control bones that were selectable.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@36367 954f8c5b-7b00-dc11-b283-0030488c597c