Fix for #28876, alpha over with empty image node gives black result.
The image node output is the default value when no image is selected. In pre-2.60 this was always initialized to 0 alpha, the defaults written in the node socket templates were completely ignored for outputs (except for value and RGB input nodes, which use these as button values). Now the stack values are initialized with the template defaults, which are all 0 by default. This patch changes alpha to 0 for image and render layer outputs too.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@40955 954f8c5b-7b00-dc11-b283-0030488c597c