Split off 3D Cursor settings from "View" panel in 3D View
This means that you don't need to have the view panel open all the
time, reducing the amount of scrolling required just to be able to
precisely position the 3D cursor at specific coordinates while
changing some other settings. While most of the settings in the View
panel are less likely to be frequently changed, the 3D cursor can in
some workflows end up needing to be accessed quite frequently.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@41575 954f8c5b-7b00-dc11-b283-0030488c597c