Fix for toggle buttons in node headers.
The buttons for "hiding" (collapsing) a node, hiding unlinked sockets, additional options, the preview and for opening a node group were all using a custom mouse test function, which was broken. They now use actual buttons instead of just displaying icons. Before executing the respective operators the button's node has to be selected and activated, so the buttons use an intermediate handle function, which selects the node and then calls the operator.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@42705 954f8c5b-7b00-dc11-b283-0030488c597c