Fix #30094: Subdivision Surfaces Modifier corrupts Copy Location
Issue was caused because of missed MDeformVert layer on subdivided cube.
If regular object is constrained to mesh object, extra customdata mask
is adding to such target. But armature bones are handling separately
and there was no such thing happens.
Ideally this code should be deduplicated to avoid such kind of issues,
but better not to do it so close before release.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@43951 954f8c5b-7b00-dc11-b283-0030488c597c