report [#30814] Absolute Shape Keys not working in 2.6
This report points out thet absolute shape keys are unusable.

The problem is there was no way to adjust the play time of a shape key (all absolte shape keys would start at frame zero with no way to change the speed).

Added an 'eval_time' property to the key block that works like the curve path evaluation time, so the time in the keyblock can be controlled.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@45409 954f8c5b-7b00-dc11-b283-0030488c597c