Renamed "fake" OpenGL identifiers.
Any identifier that looks like an OpenGL identifier, but isn't, causes a false alarm by the glreport.py tool.  Most of these were in comments so I just rephrased the comments.  There were a couple of static functions/macros that were easy enough to rename.  Only the glTexco and glIndex fields of the DMVertexAttribs struct was public and had non-local uses.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@46270 954f8c5b-7b00-dc11-b283-0030488c597c