Bugfix [#27886] Transform constraint maps wrongly with negative scale
AFAIK, it is impossible to determine exactly which axes may have negative
scaling values from a 4x4 matrix (which is the underlying cause of this bug).
However, we can figure out if there is some negative scaling going on in that
matrix (i.e. one of the axes has negative scale). So, the fix here is to
negatively scale everything if we detect this happening.

WARNING: do not rely on being able to accurately detecting positive/negative
values for more than a single axis per bone controller. Weird results may occur.
You have been warned.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@47721 954f8c5b-7b00-dc11-b283-0030488c597c