Fix #32058, Crash when using ParticleInstance with an hidden particle system.
The instance modifier needs to access the derived mesh data of the particle parent object to create stuff on the hairs, however the dm does not exist when the particle modifier is hidden. This is a general design problem: Objects accessing another object's derived mesh data is unsafe. For now it just checks valid dm pointer and uses identity transform if NULL.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@48837 954f8c5b-7b00-dc11-b283-0030488c597c