Fix #31339, Modifying mesh destroys particle system.
The particle system modifier has to ensure tesselation before testing for topology changes. It compares the number of vertices, edges and tesselation faces to the previously stored values.

Note that this test only detects a subset of actual topology changes (where the number of elements differs), but this is a known limitation we have to live with for now.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@49199 954f8c5b-7b00-dc11-b283-0030488c597c