Fix #32139: Making vertex parent hides faces
Call normals update and re-tesselate the BMesh before generating
derived mesh for vertex parent. this is needed for proper display
of mesh in edit mode.

Tried to use EDBM_update_generic, but it gave artifacts due to it
doesn't update normals. usually it's not a problem, because it's used
at the end of operator and all needed data is handles by depsgraph.
It doesn't work for vertex parent because derived mesh is being
created here outside of generic object update, so one extra manual
step seems to be needed here.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@49233 954f8c5b-7b00-dc11-b283-0030488c597c