Fix for incomplete Reroute node updates. Adding reroute nodes by spliting links would often result in unrelated color sockets or otherwise miss updates.
The reason is that the per-node updates used for Reroute node type inheritance are not supposed to be looking at connected nodes, they are purely for "local" updates. For this sort of "global" update which requires depth-first search, the update function on the node tree level must be used instead.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@49630 954f8c5b-7b00-dc11-b283-0030488c597c