Bugfix [#32331] Graph editor zoom to selected includes coordinate origin if more than one curve is displayed
The range calculation used to use a fixed 0-1 range whenever it couldn't find
any values for a particular F-Curve. However, this was then taken by the
aggregation calculation to be used as just another value, leading to problems if
only vertices of a very high-value curve are selected to be included.

Modified the range calculation to ensure that suitable vertices were found
before trying to take the range values returned.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@49976 954f8c5b-7b00-dc11-b283-0030488c597c