calculate sticky wasnt working very well and had a few glites - updating data was only being done on the active object but sticly was being calculated for the selection.
split this into 2 operators, one that works on the selection and another that operates on the active object - so we can have a button in the mesh panels that calculates sticky.

also note that there was no way to calculate sticky from the UI - perhaps this feature should die a quiet death?

anyway - it works better then it used to for now.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@50787 954f8c5b-7b00-dc11-b283-0030488c597c