Bugfix [#32647] PolyLine tool for Grease Pencil was broken
This was broken in r.46589, although it doesn't seem that these changes have any
relevance to the main fix being performed there.

The problem was that the offending changes made the Grease Pencil modal handler
exit when an RMB event occurs meant that the operator was exiting after the
initial click, meaning that it was only possible to draw single dots at a time
when using the hotkey version of PolyLine. The toolbox version however was
unaffected.

I've noted this specific problem in the code as a warning.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@50885 954f8c5b-7b00-dc11-b283-0030488c597c