Parenting an object to a deformer (armature/curve/lattice) will now attempt to check if the object is already parented to said deformer before trying to add a new modifier
This should help reduce the number of cases where users inadvertantly end up
creating multiple deform modifiers pointing to the same object, which has been
known to be a cause of "double-transform" artifacts.

Note that this is only able to detect these cases by checking if the parent
object is selected, so this will only really work for the Ctrl-P shortcut where
you have to select both objects first. However, it shouldn't be a problem either
in the Outliner (drag and drop), as the object probably won't be a child of its
parent already if you're doing this.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@51327 954f8c5b-7b00-dc11-b283-0030488c597c