Bugfix #24343
Ancient 2.5 UI issue:

The button views in Property Editor have an internal storage, to ensure the view
on the buttons stays same when you switch contextes or select objects.

Bug was that - in case buttons were all scrolled invisible - sometimes no slider
was drawn to indicate that.

Discussion on whether this is a good feature or not: the only improvement would
be to store such settings even per object... but that's a feature creep I think.




git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@51626 954f8c5b-7b00-dc11-b283-0030488c597c