Bugfix #28734
Double clicks were never working reliably in Blender - this mostly because
it wasn't a real event, but something generated in the handler code.

Now it is an actual event - meaning it always gets handled (if you have
a keymap item for it of course), but if there's no doubleclick handling
it treats the doubleclick as a normal click.

Also cleaned code. No recursion anymore.



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@51661 954f8c5b-7b00-dc11-b283-0030488c597c