fix [#33501] Grease pencil in OpenGL render
With the view3d 'Render Only' option, grease pencil wouldn't draw, but for OpenGL render it did.

Since grease pencil can be very useful in opengl renders, enable grease pencil drawing with 'Render Only' option in the viewport,
and add a checkbox in the grease pencil header not to draw (unchecking each layer is annoying and applies to all spaces).


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@53068 954f8c5b-7b00-dc11-b283-0030488c597c