UI todo:
Added "Panel Title" style to Theme settings.
Allows to make these nice larger or draw differently.

Also tried to put hinting in Style, but this needs to be a per-font
setting. uiFont data is still not being saved (also not allowing
to set own font files for UI). That's a todo for 267 then.



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@54049 954f8c5b-7b00-dc11-b283-0030488c597c