Bugfix: Selecting AnimData "expanders" in AnimEditors works again
Somewhere along the line, this functionality broke, even though the code to
handle these settings was still in place for many of these. The main implication
of this fix is that it should now be possible to select a particular AnimData
block, which makes it possible to do things such as changing the action
associated with that AnimData block (i.e. via the "Animation Data" panel in the
NLA Editor), as well as other operations which I've had on the todolist for a
while. Stay tuned!


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@55261 954f8c5b-7b00-dc11-b283-0030488c597c