Fix #34697, Undoing a node group crashes Blender.
The problem here was that the extra forward compatibility data for node groups (links to the node interface sockets) was written for files as well as undo memfiles. But only in the case of true .blend files is this data subsequently removed again after loading, which in the case of undo leaves invalid links hanging around in the node tree. Forward compat data is now skipped for undo.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@55426 954f8c5b-7b00-dc11-b283-0030488c597c