BGE bugfix: [#34550] getVertexArrayLength changed in 2.66
This bug was producing way more geometry than we need. Potentially
memory saver (or performance even) bugfix.

Bottomline is:
We need to initialize the UVs otherwise vertex comparison fails ...
Ideally MoTo should take care of initializing MT_Point2
(note in Windows that seems to be fine, but it's safer to not rely on that)

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@55483 954f8c5b-7b00-dc11-b283-0030488c597c