UI drawing fix:
Wrongly replaced a "&" with "|" for a define. That made node headers in node editor
draw too dark. Also made header for default node not use alpha.

In general this drawing is not very good - it gets too transparent and dark.
Needs nicer AA function in interface, will be done later.



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@56170 954f8c5b-7b00-dc11-b283-0030488c597c