Reconstructed scene scale ambiguity improvement
Made it so reconstructed scene always scaled in a way
that variance of camera centers is unity.

This solves "issues" when different keyframes will
give the same reprojection error but will give scenes
with different.scale, which could easily have been
considered as a bad keyframe combination.

This change is essential for automatic keyframe
selection algorithm to work reliable for user.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@56632 954f8c5b-7b00-dc11-b283-0030488c597c