BGE: fix [#32775] LibFree crashing
Used the engine's RemoveScene (which ultimately calls the converter's RemoveScene) instead of the converter's RemoveScene when the converter free a blend file. This handles the scene removal in a safer fashion and solves the crash described in part 2 of the bug report. Part 1 no longer appears to cause a crash.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@57626 954f8c5b-7b00-dc11-b283-0030488c597c