BGE: Adding partial support for LibLoaded lights with GLSL materials.
Any GLSL materials loaded after lights are LibLoaded will now use the lights in
heir shaders. This includes materials loaded from the same scene as the LibLoaded
lights. We could later add a new flag to LibLoad to recompile all existing shaders,
but this commit should offer a lot more flexibility as is.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@59196 954f8c5b-7b00-dc11-b283-0030488c597c