Use reentrant qsort() in particle codes
Particle system code used global variable to sort hair by orig index,
which is not safe for threading at all.

Replaced this with usage of reentrant version of qsort, which is
now implemented in BLI. It was moved from recast navigation code
to BLI, so more areas could use it (if needed).

--
svn merge -r59086:59087 ^/branches/soc-2013-depsgraph_mt



git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@59279 954f8c5b-7b00-dc11-b283-0030488c597c