Image Editor: implement FKey to call 'View All' with 'fit_view' This mimics the behaviour we have in the Clip Editor.
I personally would prefer if we had no border once in fullscreen
(current border is 5 pixels).

I will consult Sergey Sharybin first to see if we can change that in the clip editor as well (though there I
believe the border is useful - the bottom of the editor is used to indicate 'tracked' frames.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@59339 954f8c5b-7b00-dc11-b283-0030488c597c