fix for window join action being interpreted as a split.
The limit for dragging a gesture on an area corner was smaller then the area hot-spot,
so you could click on the right-most side of the bottom-left corner, drag left - pass the gesture threshold and still be in the same area. so a motion intended as a join would register as a split.
Happened more with high DPI values. fix by ensuring the drag limit is always higher then the hotspot.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@59794 954f8c5b-7b00-dc11-b283-0030488c597c