Fix for [#36707] Blender Opens in fullscreen and stays like that Reported by holy enigma
The previous commit in this area removed bounding box checks, because
they were done against primary monitor. Now do bound checks against
the entire desktop, or rather, the virtual screen. This is the bounding
rectangle of all the monitors.

This should ensure windows are always created within the confines of this
area.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@60067 954f8c5b-7b00-dc11-b283-0030488c597c