Fix #36718: Wrong lighting on text objects
In fact, all curve objects were flipped in GLSL mode.
This is because of the way how normals are calculated
for them (inwards vs. outwards).

We might want to make normals consistent all over, but
that would be the bigger change. For ow just made some
tweaks to OpenGL setup.

Thanks Brecht for the review and tests!

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@60201 954f8c5b-7b00-dc11-b283-0030488c597c