Fix: ensure cycles mist pass stays in range 0..1, it could have values out of this range due to sampling noise.
Side note: I looked into the mist pass because it was apparently not calculating
mist correctly on characters with transparent hair. Turns out this is just
sampling noise that goes away with more samples.

This noise is because the ray will randomly go to the next transparency layer or
get reflected, the path tracing integrator will not branch the path and only pick
one of the two directions each time.

Branched path tracing however will shade all transparent layers for each AA
sample, which means this source of noise is eliminated.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@60853 954f8c5b-7b00-dc11-b283-0030488c597c