Fix #37261 Rendering a Render Layer from another scene doesn't update.
The scene pointer used for looking up the appropriate source of render result images in the image editor was always taken from context. This means that render results for a different scene would never be
displayed in the image editor.

To give feedback on running renders, try to get the running render job's scene pointer in the image editor for render result type images. This only happens during rendering, apart from that the regular
context scene result is displayed.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@61033 954f8c5b-7b00-dc11-b283-0030488c597c